Release 0.5.0

+++ android/src/main/java/changjoopark/com/flutter_foreground_plugin/FlutterForegroundPlugin.java
@@ -47,27 +47,27 @@ public class FlutterForegroundPlugin implements MethodCallHandler {
 
     @Override
     public void onMethodCall(MethodCall call, Result result) {
-        switch (call.method) {
-            case "startForegroundService":
-                final Boolean holdWakeLock = call.argument("holdWakeLock");
-                final String icon = call.argument("icon");
-                final int color = ((Long)call.argument("color")).intValue();
-                final String title = call.argument("title");
-                final String content = call.argument("content");
-                final String subtext = call.argument("subtext");
-                final Boolean chronometer = call.argument("chronometer");
-                final Boolean stopAction = call.argument("stop_action");
-                final String stopIcon = call.argument("stop_icon");
-                final String stopText = call.argument("stop_text");
+                switch (call.method) {
+                    case "startForegroundService":
+                        final Boolean holdWakeLock = call.argument("holdWakeLock");
+                        final String icon = call.argument("icon");
+                        final int color = call.argument("color");
+                        final String title = call.argument("title");
+                        final String content = call.argument("content");
+                        final String subtext = call.argument("subtext");
+                        final Boolean chronometer = call.argument("chronometer");
+                        final Boolean stopAction = call.argument("stop_action");
+                        final String stopIcon = call.argument("stop_icon");
+                        final String stopText = call.argument("stop_text");
 
-                launchForegroundService(icon, color, title, content, subtext, chronometer, stopAction, stopIcon, stopText);
-                result.success("startForegroundService");
-                break;
-            case "stopForegroundService":
-                stopForegroundService();
-                result.success("stopForegroundService");
-                break;
-            case "setServiceMethodInterval":
+                        launchForegroundService(icon, color, title, content, subtext, chronometer, stopAction, stopIcon, stopText);
+                        result.success("startForegroundService");
+                        break;
+                    case "stopForegroundService":
+                        stopForegroundService();
+                        result.success("stopForegroundService");
+                        break;
+                    case "setServiceMethodInterval":
                 if (call.argument("seconds") == null) {
                     result.notImplemented();
                     break;
